Added DLLReloaderUnity for auto dll loading
DLLReloaderUnity automatically unloads dlls on editor unfocus and loads them when you focus the editor. This lets you:
1. Create symlink to robot sim dll in Assets/Plugins
2. Build whenever you switch out of unity without file errors

- Created VexSimulator namespace

Was going to implement this in c++, that is the reason for the DLLReloader.cs. But it was buggy and for some reason did not work with unity. I'll leave the cs file here for now in case I want to come back to it. However, the current solution works just fine.